Created alternative post location.
