+++ client/src/App.js
@@ -3,7 +3,6 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Header from "./components/Header";
 import Home from "./pages/Home";
 import PostPage from "./pages/PostPage";
-import CreatePost from "./pages/CreatePost";
 
 function App() {
   return (
@@ -12,7 +11,6 @@ function App() {
       <Router>
         <Routes>
           <Route path="/post-page" element={<PostPage />} />
-          <Route path="/create-post" element={<CreatePost />} />
           <Route path="/" element={<Home />} />
         </Routes>
       </Router>
