--- client/src/components/PostButton.js
@@ -1,9 +1,0 @@
-const PostButton = () => {
-  return (
-    <button className="transition ease-in-out duration-300 text-white bg-lightblue hover:bg-blue text-xl py-2 rounded-md w-full ">
-      Create a Post!
-    </button>
-  );
-};
-
-export default PostButton;
